requirements/dev.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
cfgv==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
cryptography==42.0.5 ; python_version >= "3.12" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
distlib==0.3.8 ; python_version >= "3.12" and python_version < "4.0"
django-redis==5.4.0 ; python_version >= "3.12" and python_version < "4.0"
django-templated-mail==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
django==5.0.4 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
djoser==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
drf-spectacular==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.13.4 ; python_version >= "3.12" and python_version < "4.0"
identify==2.5.36 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.12" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.2.1 ; python_version >= "3.12" and python_version < "4.0"
pre-commit==3.7.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyjwt==2.8.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
redis==5.0.4 ; python_version >= "3.12" and python_version < "4.0"
referencing==0.35.0 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.12" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.12" and python_version < "4.0"
setuptools==69.5.1 ; python_version >= "3.12" and python_version < "4.0"
social-auth-app-django==5.4.1 ; python_version >= "3.12" and python_version < "4.0"
social-auth-core==4.5.4 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
virtualenv==20.26.0 ; python_version >= "3.12" and python_version < "4.0"
